@@ -1,10 +1,10 @@
 name: timeline_tile_plus
 description: A package to help you build highly customisable timelines with Flutter.
-version: 0.0.3
+version: 0.1.0
 homepage: https://github.com/dtilhoo/timeline_tile_plus
 
 environment:
-  sdk: ^3.5.3
+  sdk: ^3.10.0
   flutter: ">=1.17.0"
 
 dependencies:
@@ -14,6 +14,6 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^4.0.0
-  
+  flutter_lints: ^5.0.0
+
 flutter:
