@@ -121,7 +121,7 @@ class _Child extends StatelessWidget {
 );
 
 class Example8Vertical extends StatelessWidget {
-  const Example8Vertical({Key? key}) : super(key: key);
+  const Example8Vertical({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -203,7 +203,7 @@ class Example8Vertical extends StatelessWidget {
 }
 
 class Example8Horizontal extends StatelessWidget {
-  const Example8Horizontal({Key? key}) : super(key: key);
+  const Example8Horizontal({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -288,10 +288,9 @@ class Example8Horizontal extends StatelessWidget {
 
 class _Child extends StatelessWidget {
   const _Child({
-    Key? key,
     required this.text,
     this.font = 'Shrikhand',
-  }) : super(key: key);
+  });
 
   final String text;
   final String font;
