@@ -12,8 +12,8 @@ class IndicatorStyle {
     this.iconStyle,
     this.indicatorXY = 0.5,
     this.drawGap = false,
-  })  : assert(width >= 0, 'The width must be provided and bigger than 0.0'),
-        assert(height >= 0, 'The height must be provided and bigger than 0.0');
+  }) : assert(width >= 0, 'The width must be provided and bigger than 0.0'),
+       assert(height >= 0, 'The height must be provided and bigger than 0.0');
 
   /// The width from the indicator.
   /// It defaults to 20.
@@ -61,11 +61,7 @@ class IndicatorStyle {
 
 /// Used to customize the icon used with the default indicator.
 class IconStyle {
-  IconStyle({
-    required this.iconData,
-    this.color = Colors.black,
-    this.fontSize,
-  });
+  IconStyle({required this.iconData, this.color = Colors.black, this.fontSize});
 
   /// The icon to render.
   final IconData iconData;
@@ -83,6 +79,12 @@ class LineStyle {
   const LineStyle({
     this.color = Colors.grey,
     this.thickness = 4,
+    this.dotted = false,
+    this.dotRadius = 4.0,
+    this.dotSpacing = 10.0,
+    this.isDashed = false,
+    this.dashLength = 4.0,
+    this.dashSpacing = 4.0,
   });
 
   /// The color used to paint the line. It defaults to ([Colors.grey]).
@@ -91,4 +93,11 @@ class LineStyle {
   /// The thickness from the line. It can't be bigger than ([IndicatorStyle.width])
   /// and defaults to 4.
   final double thickness;
+  final bool dotted; // Flag to indicate dotted line style
+  final double dotRadius; // Length of each dash
+  final double dotSpacing; // Space between dashes
+
+  final bool isDashed;
+  final double dashLength;
+  final double dashSpacing;
 }
